chore: tab-provider 삭제 및 구조 변경
