+++ packages/bds-ui/src/components/index.ts
@@ -1,4 +1,3 @@
 export { default as Button } from './button/button';
 export { default as Tab } from './tab/tab';
-export { default as TabProvider } from './tab/tab-provider';
 export { default as ThemeProvider } from './theme-provider';
